+++ src/argus/notificationprofile/serializers.py
@@ -155,10 +155,16 @@ class RequestDestinationConfigSerializer(serializers.ModelSerializer):
         if self.instance and "media" in attrs.keys() and not attrs["media"].slug == self.instance.media.slug:
             raise serializers.ValidationError("Media cannot be updated, only settings.")
         if "settings" in attrs.keys():
+            if type(attrs["settings"]) != dict:
+                raise serializers.ValidationError("Settings has to be a dictionary.")
             if self.instance:
-                attrs["settings"] = MEDIA_CLASSES_DICT[self.instance.media.slug].validate(self, attrs)
+                attrs["settings"] = MEDIA_CLASSES_DICT[self.instance.media.slug].validate(
+                    self, attrs, self.context["request"].user
+                )
             else:
-                attrs["settings"] = MEDIA_CLASSES_DICT[attrs["media"].slug].validate(self, attrs)
+                attrs["settings"] = MEDIA_CLASSES_DICT[attrs["media"].slug].validate(
+                    self, attrs, self.context["request"].user
+                )
 
         return attrs